+++ bin/integrated_app/config_models.py
@@ -444,11 +444,11 @@ def resolve_model_path(
 
     shared 模式:
         {comfy_models_dir}/{mount_map[sub_dir]}/{sub_path}
-        例如: C:/Users/Doro/APP/ComfyUI/models/text_encoders/FLUX.2-klein-9b/qwen_3_8b_fp8mixed.safetensors
+        例如: C:/Users/Doro/APP/ComfyUI/models/text_encoders/Z_image(turbo)/qwen_3_4b_fp8_mixed.safetensors
 
     portable 模式:
         {project_root}/{internal_models_dir}/{sub_dirs[sub_dir]}/{sub_path}
-        例如: ./pretrained_models/text_encoders/FLUX.2-klein-9b/qwen_3_8b_fp8mixed.safetensors
+        例如: ./pretrained_models/text_encoders/Z_image(turbo)/qwen_3_4b_fp8_mixed.safetensors
 
     Args:
         model_paths: 引擎配置中的 text_encoder / unet / vae 等 ModelPaths